+++ core/src/main/java/com/graphhopper/routing/ev/BikeRoadAccessConditional.java
@@ -28,16 +28,16 @@ import java.util.HashSet;
  * Stores temporary so-called conditional restrictions from access:conditional and other conditional
  * tags affecting bikes.
  */
-public enum BikeAccessConditional {
+public enum BikeRoadAccessConditional {
 
     MISSING, YES, NO;
 
     public static final Collection<String> CONDITIONALS = new HashSet<>(Arrays.asList("access:conditional",
             "vehicle:conditional", "bicycle:conditional"));
-    public static final String KEY = "bike_access_conditional";
+    public static final String KEY = "bike_road_access_conditional";
 
-    public static EnumEncodedValue<BikeAccessConditional> create() {
-        return new EnumEncodedValue<>(KEY, BikeAccessConditional.class);
+    public static EnumEncodedValue<BikeRoadAccessConditional> create() {
+        return new EnumEncodedValue<>(KEY, BikeRoadAccessConditional.class);
     }
 
     @Override