+++ core/src/test/java/com/graphhopper/routing/CustomizableConditionalRestrictionsTest.java
@@ -6,8 +6,7 @@ import com.graphhopper.GraphHopper;
 import com.graphhopper.GraphHopperConfig;
 import com.graphhopper.config.Profile;
 import com.graphhopper.json.Statement;
-import com.graphhopper.routing.ev.FootAccessConditional;
-import com.graphhopper.routing.util.parsers.DefaultTagParserFactory;
+import com.graphhopper.routing.ev.FootRoadAccessConditional;
 import com.graphhopper.util.CustomModel;
 import com.graphhopper.util.Helper;
 import com.graphhopper.util.details.PathDetail;
@@ -36,7 +35,7 @@ public class CustomizableConditionalRestrictionsTest {
     public void testConditionalAccess() {
         GraphHopper hopper = new GraphHopper().
                 setStoreOnFlush(false).
-                setEncodedValuesString(FootAccessConditional.KEY);
+                setEncodedValuesString(FootRoadAccessConditional.KEY);
 
         hopper.init(new GraphHopperConfig().
                 setProfiles(Arrays.asList(new Profile("foot").setVehicle("foot"))).
@@ -47,7 +46,7 @@ public class CustomizableConditionalRestrictionsTest {
                 putObject("datareader.date_range_parser_day", "2023-08-01"));
         hopper.importOrLoad();
 
-        String PD_KEY = "street_access_conditional";
+        String PD_KEY = "tag_access_conditional";
         GHResponse rsp = hopper.route(new GHRequest(50.909136, 14.213924, 50.90918, 14.213549).
                 setProfile("foot").
                 setPathDetails(Arrays.asList(PD_KEY)));
@@ -59,7 +58,7 @@ public class CustomizableConditionalRestrictionsTest {
 
         rsp = hopper.route(new GHRequest(50.909136, 14.213924, 50.90918, 14.213549).
                 setProfile("foot").
-                setCustomModel(new CustomModel().addToPriority(Statement.If("foot_access_conditional == NO", Statement.Op.MULTIPLY, "0"))).
+                setCustomModel(new CustomModel().addToPriority(Statement.If("foot_road_access_conditional == NO", Statement.Op.MULTIPLY, "0"))).
                 setPathDetails(Arrays.asList(PD_KEY)));
         assertFalse(rsp.hasErrors(), rsp.getErrors().toString());
         assertEquals(16, rsp.getBest().getDistance(), 1);