+++ core/src/test/java/com/graphhopper/routing/util/parsers/OSMRoadAccessConditionalParserTest.java
@@ -2,7 +2,7 @@ package com.graphhopper.routing.util.parsers;
 
 import com.graphhopper.reader.ReaderWay;
 import com.graphhopper.routing.ev.ArrayEdgeIntAccess;
-import com.graphhopper.routing.ev.CarAccessConditional;
+import com.graphhopper.routing.ev.CarRoadAccessConditional;
 import com.graphhopper.routing.ev.EnumEncodedValue;
 import com.graphhopper.routing.util.EncodingManager;
 import com.graphhopper.storage.IntsRef;
@@ -10,30 +10,30 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class OSMConditionalRestrictionsParserTest {
+class OSMRoadAccessConditionalParserTest {
 
-    private final EnumEncodedValue<CarAccessConditional> restricted = CarAccessConditional.create();
+    private final EnumEncodedValue<CarRoadAccessConditional> restricted = CarRoadAccessConditional.create();
     private final EncodingManager em = new EncodingManager.Builder().add(restricted).build();
-    private final OSMConditionalRestrictionsParser parser = new OSMConditionalRestrictionsParser(CarAccessConditional.CONDITIONALS,
-            (edgeId, access, b) -> restricted.setEnum(false, edgeId, access, b ? CarAccessConditional.YES : CarAccessConditional.NO), "2023-05-17");
+    private final OSMRoadAccessConditionalParser parser = new OSMRoadAccessConditionalParser(CarRoadAccessConditional.CONDITIONALS,
+            (edgeId, access, b) -> restricted.setEnum(false, edgeId, access, b ? CarRoadAccessConditional.YES : CarRoadAccessConditional.NO), "2023-05-17");
 
     @Test
     public void testBasics() {
         String today = "2023 May 17";
         ArrayEdgeIntAccess edgeIntAccess = new ArrayEdgeIntAccess(1);
         int edgeId = 0;
-        assertEquals(CarAccessConditional.MISSING, restricted.getEnum(false, edgeId, edgeIntAccess));
+        assertEquals(CarRoadAccessConditional.MISSING, restricted.getEnum(false, edgeId, edgeIntAccess));
 
         ReaderWay way = new ReaderWay(0L);
         way.setTag("highway", "road");
         way.setTag("access:conditional", "no @ (" + today + ")");
         parser.handleWayTags(edgeId, edgeIntAccess, way, IntsRef.EMPTY);
-        assertEquals(CarAccessConditional.NO, restricted.getEnum(false, edgeId, edgeIntAccess));
+        assertEquals(CarRoadAccessConditional.NO, restricted.getEnum(false, edgeId, edgeIntAccess));
 
         edgeIntAccess = new ArrayEdgeIntAccess(1);
         way.setTag("access:conditional", "no @ ( 2023 Mar 23 - " + today + " )");
         parser.handleWayTags(edgeId, edgeIntAccess, way, IntsRef.EMPTY);
-        assertEquals(CarAccessConditional.NO, restricted.getEnum(false, edgeId, edgeIntAccess));
+        assertEquals(CarRoadAccessConditional.NO, restricted.getEnum(false, edgeId, edgeIntAccess));
 
         edgeIntAccess = new ArrayEdgeIntAccess(1);
         way.clearTags();
@@ -41,24 +41,24 @@ class OSMConditionalRestrictionsParserTest {
         way.setTag("access", "no");
         way.setTag("access:conditional", "yes @ (" + today + ")");
         parser.handleWayTags(edgeId, edgeIntAccess, way, IntsRef.EMPTY);
-        assertEquals(CarAccessConditional.YES, restricted.getEnum(false, edgeId, edgeIntAccess));
+        assertEquals(CarRoadAccessConditional.YES, restricted.getEnum(false, edgeId, edgeIntAccess));
 
         // range does not match => missing
         edgeIntAccess = new ArrayEdgeIntAccess(1);
         way.setTag("access:conditional", "no @ ( 2023 Mar 23 )");
         parser.handleWayTags(edgeId, edgeIntAccess, way, IntsRef.EMPTY);
-        assertEquals(CarAccessConditional.MISSING, restricted.getEnum(false, edgeId, edgeIntAccess));
+        assertEquals(CarRoadAccessConditional.MISSING, restricted.getEnum(false, edgeId, edgeIntAccess));
 
         // for now consider if seasonal range
         edgeIntAccess = new ArrayEdgeIntAccess(1);
         way.setTag("access:conditional", "no @ ( Mar 23 - Aug 23 )");
         parser.handleWayTags(edgeId, edgeIntAccess, way, IntsRef.EMPTY);
-        assertEquals(CarAccessConditional.NO, restricted.getEnum(false, edgeId, edgeIntAccess));
+        assertEquals(CarRoadAccessConditional.NO, restricted.getEnum(false, edgeId, edgeIntAccess));
 
         edgeIntAccess = new ArrayEdgeIntAccess(1);
         way.setTag("access:conditional", "yes @ Apr-Nov");
         parser.handleWayTags(edgeId, edgeIntAccess, way, IntsRef.EMPTY);
-        assertEquals(CarAccessConditional.YES, restricted.getEnum(false, edgeId, edgeIntAccess));
+        assertEquals(CarRoadAccessConditional.YES, restricted.getEnum(false, edgeId, edgeIntAccess));
     }
 
     @Test
@@ -70,11 +70,11 @@ class OSMConditionalRestrictionsParserTest {
         // ignore incomplete values
         way.setTag("access:conditional", "no @ 2023 Mar-Oct");
         parser.handleWayTags(edgeId, edgeIntAccess, way, IntsRef.EMPTY);
-        assertEquals(CarAccessConditional.MISSING, restricted.getEnum(false, edgeId, edgeIntAccess));
+        assertEquals(CarRoadAccessConditional.MISSING, restricted.getEnum(false, edgeId, edgeIntAccess));
 
         // here the "1" will be interpreted as year -> incorrect range
         way.setTag("access:conditional", "no @ 1 Nov - 1 Mar");
         parser.handleWayTags(edgeId, edgeIntAccess, way, IntsRef.EMPTY);
-        assertEquals(CarAccessConditional.MISSING, restricted.getEnum(false, edgeId, edgeIntAccess));
+        assertEquals(CarRoadAccessConditional.MISSING, restricted.getEnum(false, edgeId, edgeIntAccess));
     }
 }